CarImages-GUID-FileHelper

+++ Entities/DTOs/CarDetailDto.cs
@@ -12,7 +12,7 @@ namespace Entities.DTOs
         public string BrandName { get; set; }
         public int ColorId { get; set; }
         public string ColorName { get; set; }
-        public int DailyPrice { get; set; }
+        public decimal DailyPrice { get; set; }
         public string ModelYear { get; set; }
         public string Description { get; set; }
